Amend cabal file
Still trying to learn.

In the test-suite section, removed the 'other-modules' line which I had
been using (when I was building an executable) to list its modules and
allow the tests to compile.  Replaced that with the project name now that
it is a library.

This works fine, but compiling the test suite gives me a missing-home-modules
warning.  Why?  I don't see what I have to worry about here because the
tests run just fine.